Renamed clicast.ts ... per Igot suggestions
