+++ src/vcastUtilities.ts
